feat: add threat score value object to standardize normalization logic
Adds a new value object to standardize normalization logic of threat score that some services can return, this also includes some changes to database, infrastructure, simulated services and real services.

+++ src/Openlysis.Domain/Common/Aggregates/MultiAnalysis.cs
@@ -42,7 +42,7 @@ public abstract class MultiAnalysis<TServiceAnalysis>
     /// <summary>
     /// Gets or sets the average threat score of the analysis.
     /// </summary>
-    public float? AverageThreatScore { get; protected set; }
+    public int? AverageThreatScore { get; protected set; }
 
     /// <summary>
     /// Gets the set of data hashes associated with the analysis.
